Tournament single elimination matchmaker
